Implement everything so that factorial works
