+++ run_wasm.js
@@ -1,0 +1,16 @@
+const fs = require("fs");
+
+// file is an array of bytes.
+const file_name = process.argv[2];
+console.log(`file: ${file_name}`);
+const file = fs.readFileSync(`./${file_name}`);
+const bytes = new Uint8Array(file);
+
+// console.log(bytes);
+
+WebAssembly.instantiate(bytes).then(({ instance }) => {
+  const fct = instance.exports.fct;
+
+  console.log(fct(6));
+})
+
