+++ tmp_repl.txt
@@ -20,13 +20,18 @@ FILE=test
 rm $TMP_WASM/$FILE.wasm
 cargo run --bin compiler
 
-
 wasm2wat $TMP_WASM/$FILE.wasm -o $TMP_WAT/$FILE.wat
+
+cat $TMP_WAT/$FILE.wat
+
+
+node run_wasm.js $TMP_WASM/$FILE.wasm
 
 wasm-objdump $TMP_WASM/$FILE.wasm -x
 
 wasm-objdump $TMP_WASM/$FILE.wasm -s
 
+wasm-objdump $TMP_WASM/$FILE.wasm -d
 
 
 # ===WAT tests===
