+++ wasm/src/binary_format/mod.rs
@@ -1,5 +1,6 @@
 pub mod primitives;
 pub mod sections;
+pub mod instructions;
 
 mod export;
 mod import;
